schema changes and create repo with commit
Former-commit-id: 36ac501998956559324532e903e7b868c8158b27

+++ catalog/cataloger_diff.go
@@ -35,7 +35,7 @@ func (c *cataloger) Diff(ctx context.Context, repository string, leftBranch stri
 	return differences.(Differences), nil
 }
 
-func (c *cataloger) doDiff(tx db.Tx, leftID, rightID int) (Differences, error) {
+func (c *cataloger) doDiff(tx db.Tx, leftID, rightID int64) (Differences, error) {
 	relation, err := getBranchesRelationType(tx, leftID, rightID)
 	if err != nil {
 		return nil, err
@@ -43,7 +43,7 @@ func (c *cataloger) doDiff(tx db.Tx, leftID, rightID int) (Differences, error) {
 	return c.doDiffByRelation(tx, relation, leftID, rightID)
 }
 
-func (c *cataloger) doDiffByRelation(tx db.Tx, relation RelationType, leftID, rightID int) (Differences, error) {
+func (c *cataloger) doDiffByRelation(tx db.Tx, relation RelationType, leftID, rightID int64) (Differences, error) {
 	switch relation {
 	case RelationTypeFromFather:
 		return c.diffFromFather(tx, leftID, rightID)
@@ -56,7 +56,7 @@ func (c *cataloger) doDiffByRelation(tx db.Tx, relation RelationType, leftID, ri
 	}
 }
 
-func (c *cataloger) diffFromFather(tx db.Tx, leftID, rightID int) (Differences, error) {
+func (c *cataloger) diffFromFather(tx db.Tx, leftID, rightID int64) (Differences, error) {
 	// get the last son commit number of the last father merge
 	// if there is none - then it is  the first merge
 	var maxSonMerge int
@@ -119,7 +119,7 @@ func diffReadDifferences(tx db.Tx) (Differences, error) {
 	return result, nil
 }
 
-func (c *cataloger) diffFromSon(tx db.Tx, leftID, rightID int) (Differences, error) {
+func (c *cataloger) diffFromSon(tx db.Tx, leftID, rightID int64) (Differences, error) {
 	// read last merge commit numbers from commit table
 	// if it is the first son-to-father commit, than those commit numbers are calculated as follows:
 	// the son is 0, as any change in the some was never merged to the father.
@@ -191,6 +191,6 @@ func (c *cataloger) diffFromSon(tx db.Tx, leftID, rightID int) (Differences, err
 	return diffReadDifferences(tx)
 }
 
-func (c *cataloger) diffNonDirect(tx db.Tx, leftID, rightID int) (Differences, error) {
+func (c *cataloger) diffNonDirect(tx db.Tx, leftID, rightID int64) (Differences, error) {
 	panic("not implemented - Someday is not a day of the week")
 }
